pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "pylogic"
version = "0.1.0"
[tool.poetry.dependencies]
lark-parser = "^0.8.2"
python = "^3.8"
[tool.poetry.dev-dependencies]
